better switch based on state of State selected
